Adds the --resume argument to triage_ansible.py (#703)
If used, each issue number will be recorded to a json file in the cachedir
when processing begins. If the bot is stopped or dies, on the next run,
it will read the json file, decrement the number and set args.start_at
This effectively resumes processing and skips the issue that is broken,
and gives a smoother experience when combined with a systemd autorestart.