Add a component testing framework to check for idempotent operation (#1260)
This test is more like an integration test that is being run through pytest. The mocks are numerous so that the entire codebase can execute and NOT call out to github, and this uses a "mock manager" context manager to setup all the mocks at start and tear them down on stop. The IssueDatabase is a github mock which has it's own abstractions to create on-demand fixtures from simple inputs.