Upgrade constraints
cffi <1.12.2 fails to install on Python >3.8. At a minimum, I need that constraint updated.
I went ahead and upgrade all the constraints while I was at it. I'm not sure if I did it properly, though.
I wasn't able to figure out how to get the 'via -r [file or URL]' comments to not show up. Using
the '--no-annotate' option removes all the annotations, which it seems were in the file originally

This is my first time using pip-compile so I could use some advice on how to use it properly.