triagers/ansible: un-hardcode ansible_core_team
Maintainers and committers may be different across different
repositories. As such, we can't hardcode the organization name "ansible"
or the teams so we're pulling it back into a configuration parameter instead.

Since the maintainers are not exclusively ansible core maintainers,
we're also renaming the property to something a bit more generic.